+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All Notable changes to `mattermost-webhook` will be documented in this file
 
+## 2.1.0 - TBD
+
+### Added
+
+`Message::createFromArray`
+`Message::__set_state`
+`Attachment::createFromArray`
+`Attachment::__set_state`
+
+### Fixed
+
+Message::attachment property visibility
+
+### Deprecated
+
+- None
+
+### Removed
+
+- None
+
 ## 2.0.0 - 2018-01-12
 
 - Initial release (drop support for PHP5)